+++ source/renderer/app/components/wallet/wallet-import/WalletSelectImportDialog.js
@@ -76,6 +76,12 @@ const messages = defineMessages({
     defaultMessage: '!!!Enter a wallet name first',
     description: 'Enter a wallet name first',
   },
+  maxWalletsReachedTooltip: {
+    id: 'wallet.select.import.dialog.maxWalletsReachedTooltip',
+    defaultMessage:
+      '!!!Daedalus supports up to 30 wallets. You will need to remove another wallet before you can import this one.',
+    description: 'Max number of wallets reached',
+  },
   walletImported: {
     id: 'wallet.select.import.dialog.walletImported',
     defaultMessage: '!!!Wallet imported',
@@ -114,6 +120,7 @@ type Props = {
   onClose: Function,
   onOpenExternalLink: Function,
   nameValidator: Function,
+  isMaxNumberOfWalletsReached: boolean,
 };
 
 @observer
@@ -145,15 +152,45 @@ export default class WalletSelectImportDialog extends Component<Props> {
     return walletStatus;
   };
 
-  getWalletStatusIcon = (wallet: ExportedByronWallet) => {
-    const { nameValidator, onToggleWalletImportSelection } = this.props;
+  getWalletStatusIcon = (wallet: ExportedByronWallet, index: number) => {
+    const {
+      nameValidator,
+      onToggleWalletImportSelection,
+      isMaxNumberOfWalletsReached,
+    } = this.props;
     let statusIcon;
     if (
       wallet.import.status === WalletImportStatuses.UNSTARTED ||
       wallet.import.status === WalletImportStatuses.PENDING ||
       wallet.import.status === WalletImportStatuses.ERRORED
     ) {
-      const disabled = wallet.name === null || !nameValidator(wallet.name);
+      const invalidWalletName =
+        wallet.name === null || !nameValidator(wallet.name);
+      const walletNotSelectable =
+        isMaxNumberOfWalletsReached &&
+        wallet.import.status !== WalletImportStatuses.PENDING;
+      const disabled = invalidWalletName || walletNotSelectable;
+
+      let isOpeningUpward = true;
+      const checkboxes = document.getElementsByClassName(
+        'SimpleCheckbox_check'
+      );
+      const topWrapper = document.getElementsByClassName(
+        'WalletSelectImportDialog_topWrapper'
+      );
+      if (checkboxes.length && wallet.hasName) {
+        const checkboxTopOffset = checkboxes[index].getBoundingClientRect().top;
+        const topWrapperTopOffset = topWrapper[0].getBoundingClientRect().top;
+        const topPart = topWrapperTopOffset + 121;
+        const spaceForTooltip = checkboxTopOffset - topPart;
+        if (
+          (walletNotSelectable && spaceForTooltip < 83) ||
+          (invalidWalletName && spaceForTooltip < 27)
+        ) {
+          isOpeningUpward = false;
+        }
+      }
+
       statusIcon = (
         <Checkbox
           onChange={() => {
@@ -167,11 +204,19 @@ export default class WalletSelectImportDialog extends Component<Props> {
       if (disabled) {
         statusIcon = (
           <Tooltip
-            className={styles.enterWalletNameTooltip}
+            className={
+              walletNotSelectable
+                ? styles.maxWalletsReachedTooltip
+                : styles.enterWalletNameTooltip
+            }
             skin={TooltipSkin}
             tip={this.context.intl.formatMessage(
-              messages.enterWalletNameTooltip
+              invalidWalletName
+                ? messages.enterWalletNameTooltip
+                : messages.maxWalletsReachedTooltip
             )}
+            isBounded={walletNotSelectable}
+            isOpeningUpward={isOpeningUpward}
             arrowRelativeToTip
           >
             {statusIcon}
@@ -291,7 +336,7 @@ export default class WalletSelectImportDialog extends Component<Props> {
               <hr className={styles.separatorTop} />
             </div>
             <div className={styles.walletsContainer}>
-              {walletsWithNames.map(wallet => {
+              {walletsWithNames.map((wallet, index) => {
                 const isDuplicate = previousWalletId === wallet.id;
                 const walletRow = (
                   <div
@@ -314,7 +359,7 @@ export default class WalletSelectImportDialog extends Component<Props> {
                       {this.getWalletStatus(wallet)}
                     </div>
                     <div className={styles.walletsStatusIcon}>
-                      {this.getWalletStatusIcon(wallet)}
+                      {this.getWalletStatusIcon(wallet, index)}
                     </div>
                   </div>
                 );
@@ -334,7 +379,7 @@ export default class WalletSelectImportDialog extends Component<Props> {
                 </div>
               )}
 
-              {walletsWithoutNames.map(wallet => {
+              {walletsWithoutNames.map((wallet, index) => {
                 const isDuplicate = previousWalletId === wallet.id;
                 const walletRow = (
                   <div
@@ -382,7 +427,7 @@ export default class WalletSelectImportDialog extends Component<Props> {
                       {this.getWalletStatus(wallet)}
                     </div>
                     <div className={styles.walletsStatusIcon}>
-                      {this.getWalletStatusIcon(wallet)}
+                      {this.getWalletStatusIcon(wallet, index)}
                     </div>
                   </div>
                 );